+++ ichnaea/async/schedule.py
@@ -68,15 +68,22 @@ CELERYBEAT_SCHEDULE = {
         'options': {'expires': 15},
     },
 
+    'schedule-update-cellarea': {  # BBB
+        'task': 'ichnaea.data.tasks.scan_areas',
+        'schedule': timedelta(seconds=9),
+        'args': (100, ),
+        'options': {'expires': 15},
+    },
+
     'update-cell': {
         'task': 'ichnaea.data.tasks.update_cell',
         'schedule': timedelta(seconds=7),
         'args': (1000, ),
         'options': {'expires': 10},
     },
-    'schedule-update-cellarea': {
-        'task': 'ichnaea.data.tasks.scan_areas',
-        'schedule': timedelta(seconds=9),
+    'update-cellarea': {
+        'task': 'ichnaea.data.tasks.update_cellarea',
+        'schedule': timedelta(seconds=8),
         'args': (100, ),
         'options': {'expires': 15},
     },